+++ Source/DesignSystem/atoms/Forms/Select.tsx
@@ -52,11 +52,12 @@ export const Select = forwardRef<HTMLOptionElement, SelectProps>(({ options, onO
             <MuiSelect
                 {...field}
                 {...selectProps as MuiSelectProps}
+                required={isRequired(selectProps.required)}
                 ref={ref}
                 labelId={`${selectProps.id}-select`}
+                onOpen={onOpen}
                 value={field.value}
                 disabled={selectProps.disabled}
-                onOpen={onOpen}
                 size='small'
                 sx={{ typography: 'body2' }}
                 autoWidth