+++ Source/SelfService/Web/integrations/connection/configuration/TextCopyBox.tsx
@@ -7,15 +7,21 @@ import { useSnackbar } from 'notistack';
 
 import { Paper, Typography, SxProps } from '@mui/material';
 
-import { Button } from '@dolittle/design-system';
+import { Button, MaxWidthBlock, MaxWidthTextBlock } from '@dolittle/design-system';
 
 export type TextCopyBoxProps = {
     /**
-     * The list of instructions to display.
-     * This will be rendered as a list with default text styling, and be made available for copying as a multi-line string.
-     * For more control, use the children prop.
+     * The list of instructions to display
+     * This will be rendered as a list with default text styling wrapped in a <p>, and be made available for copying as a multi-line string
+     * For more control, use the children and instructionsToCopy props.
      */
-    instructions: string[];
+    instructions: React.ReactNode | React.ReactNode[];
+
+    /**
+     * The list of instructions to be copied to the clipboard
+     * If not supplied, the instructions prop will be used instead
+     */
+    instructionsToCopy?: string[];
 
     /**
      * Pass in children if you want to have more control over the styling of the instructions.
@@ -23,31 +29,59 @@ export type TextCopyBoxProps = {
     children?: React.ReactNode;
 
     /**
-     * Style overrides applied to the root Paper component.
+     * Should content be wrapped in a {@link MaxWidthBlock}?
+     */
+    withMaxWidth?: boolean;
+
+    /**
+     * Style overrides applied to the root Paper component
      */
     sx?: SxProps;
 };
 
-export const TextCopyBox = ({ instructions, children, sx }: TextCopyBoxProps) => {
+export const TextCopyBox = ({ instructions, instructionsToCopy, children, withMaxWidth, sx }: TextCopyBoxProps) => {
     const { enqueueSnackbar } = useSnackbar();
 
     const handleTextCopy = useCallback(() => {
-        const textToCopy = instructions.join('\n\n');
+        const textToCopy = !!instructionsToCopy ? instructionsToCopy.join('\n\n') : Array.isArray(instructions) ? instructions.join('\n\n') : instructions!.toString();
         navigator.clipboard.writeText(textToCopy);
         enqueueSnackbar('Copied to clipboard');
     }, [instructions]);
 
     return (
-        <Paper elevation={0} sx={{ 'mt': 3, 'p': 2, '& p': { mb: 3 }, ...sx }}>
-            {children
-                ? <>{children}</>
-                : <>{instructions.map((instruction, index) => <Typography key={index}>{instruction}</Typography>)}</>
-            }
-            <Button
-                label='Copy content'
-                startWithIcon='CopyAllRounded'
-                onClick={handleTextCopy}
-            />
-        </Paper>
+        <>{withMaxWidth
+            ? <MaxWidthTextBlock>
+                <TextCopyContent instructions={instructions} sx={sx} handleTextCopy={handleTextCopy}>{children}</TextCopyContent>
+            </MaxWidthTextBlock>
+            : <TextCopyContent instructions={instructions} sx={sx} handleTextCopy={handleTextCopy}>{children}</TextCopyContent>
+        }
+        </>
     );
 };
+
+type TextCopyContentProps = TextCopyBoxProps & {
+    handleTextCopy: () => void;
+};
+
+const TextCopyContent = ({ instructions, children, sx, handleTextCopy }: TextCopyContentProps) => <>
+    <Paper elevation={0} sx={{ 'mt': 3, 'p': 2, '& p': { mb: 3 }, ...sx }}>
+        <InstructionContent instructions={instructions}>{children}</InstructionContent>
+        <Button
+            label='Copy content'
+            startWithIcon='CopyAllRounded'
+            onClick={handleTextCopy}
+        />
+    </Paper>
+
+</>;
+
+type RenderContentsProps = Pick<TextCopyBoxProps, 'instructions' | 'children'>;
+
+const InstructionContent = ({ instructions, children }: RenderContentsProps) => <>
+    {children
+        ? <>{children}</>
+        : <>{Array.isArray(instructions)
+            ? instructions.map((instruction, index) => <Typography key={index}>{instruction}</Typography>)
+            : <Typography>{instructions}</Typography>
+        }</>
+    }</>;