+++ Source/SelfService/Web/integrations/connection/connectionDetails/configuration/index.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import React, { useRef, useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 
 import { useQueryClient } from '@tanstack/react-query';
 import { enqueueSnackbar } from 'notistack';
@@ -13,8 +13,8 @@ import { AccordionList, AccordionListProps, FileUploadFormRef } from '@dolittle/
 
 import { useConnectionsIdGet, useConnectionsIdDelete } from '../../../../apis/integrations/connectionsApi.hooks';
 import { CACHE_KEYS } from '../../../../apis/integrations/CacheKeys';
+import { getConnectionStatus } from '../../../../utils/helpers';
 import { useConnectionIdFromRoute } from '../../../routes.hooks';
-
 import { M3ConfigurationForm, M3ConfigurationFormRef } from '../../configuration/M3ConfigurationForm';
 import { MainM3ConnectionInfo } from '../../configuration/MainM3ConnectionInfo';
 import { useBuildConfigurationAccordionList } from '../../configuration/useBuildConfigurationAccordionList';
@@ -36,6 +36,17 @@ export const ConfigurationView = () => {
 
     const connection = query.data?.value;
     const links = query.data?.links || [];
+
+    useEffect(() => {
+        if (!connection) {
+            return;
+        }
+
+        if(getConnectionStatus(connection.status.name).label === 'pending') {
+            setEditMode(true);
+        };
+
+    }, [connection?.status.name]);
 
     const deploymentType = connection?.chosenEnvironment?.value;
     const hasSelectedDeploymentType = deploymentType?.toLowerCase() !== 'unknown';